requirements.txt
gradio
llama_index
elasticsearch
sentence-transformers
langchain-openai
faiss-cpu
langchain
langchain-community
python-docx
pymupdf
pymupdf4llm
langchain_text_splitters
bs4
nltk
pytesseract
pillow
llama-index-vector-stores-postgres
llama-index-readers-elasticsearch 
llama_index.vector_stores-elasticsearch
python-pptx
unstructured
openpyxl
